+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.10.4 - Unreleased
 ### Changed
 - Changed action URIs to relative URIs in template tags ([#345](https://github.com/putyourlightson/craft-blitz/issues/345)).
+- Element queries are no longer cached when a query contains an expression criteria ([#346](https://github.com/putyourlightson/craft-blitz/issues/346)).
 
 ## 3.10.3 - 2021-09-05
 ### Changed